refactor: move measurement code
